read_tags & show_tag_types moved main.py --> common.py
